Bring the gfx1151 attention branch current with origin/develop.

Conflicts were confined to the DSL core and were all textual:

* ir.py / lower_llvm.py -- upstream independently landed exp2_fast with the
  same math.exp2_fast op name and the same llvm.amdgcn.exp2.f32 lowering, so
  the code bodies were identical and only the docstring and one comment
  differed. Upstream's wording is kept; our duplicate is dropped.
* lower_llvm.py _Lowerer.__init__ and the trailing metadata emitter -- our
  _needs_nontemporal_md flag and its "!2 = !{i32 1}" node landed adjacent to
  upstream's _needs_av_scope_md / _debug and "!3". Both sides are kept; "!2"
  is unused upstream, so there is no metadata-id collision.

No gfx1151 kernel, instance, builder, test or integration file conflicted.

+++ projects/rocsparse/library/src/common/rocsparse_assign_async.cpp
@@ -88,7 +88,7 @@ rocsparse_status rocsparse::assign_max_async(int64_t             n,
                                              void*               dest,
                                              hipStream_t         stream)
 {
-    switch(indextype)
+    switch(static_cast<int>(indextype))
     {
     case rocsparse_indextype_i32:
     {
